Refactor: duplicated code for adapters now in seperate file
